build: Improve adherence GNU Directory Variable standard
Most UNIX systems these days standardized on Filesystem Hierarchy Standard,
so for installing user would just want to change the installation prefix,
not the individual directory variables.

For that reason, I added a PREFIX variable that is used by all the other
directory variables so user will be able to just run `make PREFIX=/usr install`,
which is much more convenient.

I also renamed the variables to follow the GNU standard [1], since I needed
to distinguish between `mandir` and `man1dir`. `man1dir` is always `$(mandir)/man1`
but `mandir` differs between Solaris and other UNIX platforms according to the Makefile.

I did not use the full granularity like `mandir` depending on `datarootdir` depending
on `prefix`, since `mandir` is the only variable that depends on `datarootdir`.
If we add another variable using it in the future, we should add `datarootdir` as well.

[1]: https://www.gnu.org/software/make/manual/html_node/Directory-Variables.html